tapes have different colors, def need to tweak though

+++ api/scripts/tape-tree.js
@@ -11,14 +11,7 @@ class TapeTree {
     )
   }
 
-  verifyProof(
-    index,
-    capacity,
-    quality,
-    style,
-    proof,
-    root
-  ) {
+  verifyProof(index, capacity, quality, style, proof, root) {
     let pair = this.toNode(index, capacity, quality, style)
     for (const item of proof) {
       pair = MerkleTree.combinedHash(pair, item)
@@ -27,10 +20,14 @@ class TapeTree {
     return pair.equals(root)
   }
 
-
   toNode(index, capacity, quality, style) {
     return Buffer.from(
-      utils.solidityKeccak256(['uint256', 'uint8', 'uint32', 'uint32'], [index, capacity, quality, style]).substr(2),
+      utils
+        .solidityKeccak256(
+          ['uint256', 'uint8', 'uint8', 'uint256'],
+          [index, capacity, quality, style]
+        )
+        .substr(2),
       'hex'
     )
   }
